Fix uint size

+++ src/game/game.cpp
@@ -7662,11 +7662,10 @@ void Game::playerCreateMarketOffer(uint32_t playerId, uint8_t type, uint16_t ite
 	}
 
 	uint64_t calcFee = (price / 100.) * amount;
-	uint32_t minFee = std::min<uint32_t>(100000, calcFee);
-	uint32_t fee = std::max<uint32_t>(20, minFee);
+	uint64_t minFee = std::min<uint64_t>(100000, calcFee);
+	uint64_t fee = std::max<uint64_t>(20, minFee);
 
 	if (type == MARKETACTION_SELL) {
-
 		if (fee > (player->getBankBalance() + player->getMoney())) {
 			return;
 		}
